Separate parser and preprocessor, add -d flag
